+++ lib/widgets/viewer/entry_viewer_stack.dart
@@ -80,7 +80,7 @@ class _EntryViewerStackState extends State<EntryViewerStack> with EntryViewContr
   late VideoActionDelegate _videoActionDelegate;
   final ValueNotifier<HeroInfo?> _heroInfoNotifier = ValueNotifier(null);
   bool _isEntryTracked = true;
-  Timer? _overlayHidingTimer;
+  Timer? _overlayHidingTimer, _videoPauseTimer;
 
   @override
   bool get isViewingImage => _currentVerticalPage.value == imagePage;
@@ -199,6 +199,7 @@ class _EntryViewerStackState extends State<EntryViewerStack> with EntryViewContr
     _verticalScrollNotifier.dispose();
     _heroInfoNotifier.dispose();
     _stopOverlayHidingTimer();
+    _stopVideoPauseTimer();
     AvesApp.lifecycleStateNotifier.removeListener(_onAppLifecycleStateChanged);
     _unregisterWidget(widget);
     super.dispose();
@@ -312,11 +313,13 @@ class _EntryViewerStackState extends State<EntryViewerStack> with EntryViewContr
   void didPushNext() => _resetSnackBarMargin();
 
   void _overrideSnackBarMargin() {
+    MarginComputer marginComputer;
     if (isViewingImage) {
-      FeedbackMixin.snackBarMarginOverrideNotifier.value = EdgeInsets.only(bottom: ViewerBottomOverlay.actionSafeHeight(context));
+      marginComputer = (context) => EdgeInsets.only(bottom: ViewerBottomOverlay.actionSafeHeight(context));
     } else {
-      FeedbackMixin.snackBarMarginOverrideNotifier.value = FeedbackMixin.snackBarMarginDefault(context);
+      marginComputer = FeedbackMixin.snackBarMarginDefault;
     }
+    FeedbackMixin.snackBarMarginOverrideNotifier.value = marginComputer;
   }
 
   void _resetSnackBarMargin() => FeedbackMixin.snackBarMarginOverrideNotifier.value = null;
@@ -333,9 +336,10 @@ class _EntryViewerStackState extends State<EntryViewerStack> with EntryViewContr
         // paused: when switching to another app
         // detached: when app is without a view
         viewerController.autopilot = false;
+        _stopVideoPauseTimer();
         pauseVideoControllers();
       case AppLifecycleState.resumed:
-        availability.onResume();
+        _stopVideoPauseTimer();
       case AppLifecycleState.hidden:
         // hidden: transient state between `inactive` and `paused`
         break;
@@ -352,9 +356,16 @@ class _EntryViewerStackState extends State<EntryViewerStack> with EntryViewContr
       // ensure playback, in case lifecycle paused/resumed events happened when switching to PiP
       await playingController?.play();
     } else {
-      await pauseVideoControllers();
+      _startVideoPauseTimer();
     }
   }
+
+  void _startVideoPauseTimer() {
+    _stopVideoPauseTimer();
+    _videoPauseTimer = Timer(ADurations.videoPauseAppInactiveDelay, pauseVideoControllers);
+  }
+
+  void _stopVideoPauseTimer() => _videoPauseTimer?.cancel();
 
   Widget _decorateOverlay(Widget overlay) {
     return ValueListenableBuilder<double>(