+++ lib/widgets/viewer/info/info_page.dart
@@ -5,6 +5,7 @@ import 'package:aves/model/entry/extensions/multipage.dart';
 import 'package:aves/model/events.dart';
 import 'package:aves/model/filters/filters.dart';
 import 'package:aves/model/selection.dart';
+import 'package:aves/model/settings/settings.dart';
 import 'package:aves/model/source/collection_lens.dart';
 import 'package:aves/theme/durations.dart';
 import 'package:aves/widgets/common/basic/insets.dart';
@@ -17,6 +18,7 @@ import 'package:aves/widgets/viewer/info/basic_section.dart';
 import 'package:aves/widgets/viewer/info/embedded/embedded_data_opener.dart';
 import 'package:aves/widgets/viewer/info/info_app_bar.dart';
 import 'package:aves/widgets/viewer/info/location_section.dart';
+import 'package:aves/widgets/viewer/info/color_section.dart';
 import 'package:aves/widgets/viewer/info/metadata/metadata_dir.dart';
 import 'package:aves/widgets/viewer/info/metadata/metadata_section.dart';
 import 'package:aves/widgets/viewer/multipage/conductor.dart';
@@ -233,10 +235,6 @@ class _InfoPageContentState extends State<_InfoPageContent> {
               ],
             ),
           );
-    final metadataSliver = MetadataSectionSliver(
-      entry: entry,
-      metadataNotifier: _metadataNotifier,
-    );
 
     return NotificationListener<FilterNotification>(
       onNotification: (notification) {
@@ -262,8 +260,16 @@ class _InfoPageContentState extends State<_InfoPageContent> {
           ),
           SliverPadding(
             padding: horizontalPadding + const EdgeInsets.only(bottom: 8),
-            sliver: metadataSliver,
+            sliver: MetadataSectionSliver(
+              entry: entry,
+              metadataNotifier: _metadataNotifier,
+            ),
           ),
+          if (!settings.useTvLayout)
+            SliverPadding(
+              padding: horizontalPadding + const EdgeInsets.only(bottom: 8),
+              sliver: ColorSectionSliver(entry: entry),
+            ),
           const BottomPaddingSliver(),
         ],
       ),