+++ CHANGELOG.md
@@ -4,6 +4,27 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+## <a id="v1.11.0"></a>[v1.11.0] - 2024-05-01
+
+### Added
+
+- Cataloguing: identify Apple variant of HDR images
+- Collection: allow using hash (md5/sha1/sha256) when bulk renaming
+- Info: color palette
+- Video: external subtitle support (SRT)
+- option to force using western arabic numerals for dates
+
+### Changed
+
+- logo
+- upgraded Flutter to stable v3.19.6
+
+### Fixed
+
+- rendering of SVG with large header
+- stopping video playback when changing device orientation on Android >=13
+- printing content orientation according to page format
+
 ## <a id="v1.10.9"></a>[v1.10.9] - 2024-04-14
 
 ### Fixed